Move into symbol package
