+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symbols/internal/impl/ast/AstSymFactory.java
@@ -7,6 +7,7 @@ package net.sourceforge.pmd.lang.java.symbols.internal.impl.ast;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
 import net.sourceforge.pmd.lang.java.ast.ASTAnyTypeDeclaration;
+import net.sourceforge.pmd.lang.java.ast.ASTCompilationUnit;
 import net.sourceforge.pmd.lang.java.ast.ASTVariableDeclaratorId;
 import net.sourceforge.pmd.lang.java.symbols.JClassSymbol;
 import net.sourceforge.pmd.lang.java.symbols.JLocalVariableSymbol;
@@ -15,12 +16,17 @@ import net.sourceforge.pmd.lang.java.symbols.JTypeParameterOwnerSymbol;
 
 public final class AstSymFactory {
 
+
+    public void createSymbolsOn(ASTCompilationUnit acu) {
+        acu.jjtAccept(AstSymbolMakerVisitor.INSTANCE, this);
+    }
+
     // keep in mind, creating a symbol sets it on the node (see constructor of AbstractAstBackedSymbol)
 
     /**
      * Builds, sets and returns the symbol for the given local variable.
      */
-    public JLocalVariableSymbol setLocalVarSymbol(ASTVariableDeclaratorId id) {
+    JLocalVariableSymbol setLocalVarSymbol(ASTVariableDeclaratorId id) {
         assert !id.isField() && !id.isEnumConstant() : "Local var symbol is not appropriate for fields";
         assert !id.isFormalParameter()
             || id.isLambdaParameter()
@@ -32,7 +38,7 @@ public final class AstSymFactory {
     /**
      * Builds, sets and returns the symbol for the given class.
      */
-    public JClassSymbol setClassSymbol(@Nullable JTypeParameterOwnerSymbol enclosing, ASTAnyTypeDeclaration klass) {
+    JClassSymbol setClassSymbol(@Nullable JTypeParameterOwnerSymbol enclosing, ASTAnyTypeDeclaration klass) {
         return new AstClassSym(klass, this, enclosing);
     }
